% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZ.R
\name{clean_names}
\alias{clean_names}
\title{Remove unwanted characters from the column names}
\usage{
clean_names(data, idf = c(), sep = "_", use.def = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{idf}{List of unwanted characters/ Identifier}

\item{sep}{Character to substitute/ Separator}

\item{use.def}{Boolean whether to use the default identifier list}
}
\value{
Dataframe with removed characters from the column names
}
\description{
Sometimes we may get a dataframe wich have column names have characters like !&()/.," "
These are unwanted obviously coz they make it impossible to reference the column using $ operator or while pipelining
So this function removes those characters from the column names.
}
\seealso{
\code{\link[base:grep]{base::gsub()}} which this function wraps.
}
