\name{ez-internal}
\alias{ez-internal}
\title{Internal ez Functions}
\description{
Internal functions for the \code{ez} package.
}
\details{
These are not to be called by the user.
}
\author{
Michael A. Lawrence \email{mike.lwrnc@gmail.com}\cr
Visit the \code{ez} development site at \url{https://github.com/mike-lawrence/ez}\cr
for the bug/issue tracker and the link to the mailing list.
}
\keyword{ internal }
