% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZ.R
\name{test_train_split}
\alias{test_train_split}
\title{Split dataset into train and test}
\usage{
test_train_split(data, train_size = 0.8, seed = 1, shuffle = FALSE)
}
\arguments{
\item{data}{Dataframes,Arrays,Vectors}

\item{train_size}{Numeric (should be between 0.0 and 1.0). Represent the proportion of the dataset to include in the test split.}

\item{seed}{Controls the shuffling applied to the data before applying the split. Pass an int for reproducible output across multiple function calls.}

\item{shuffle}{Whether or not to shuffle the data before splitting. Default is set to FALSE}
}
\value{
List containing train-test split of inputs.
}
\description{
Splitting dataset into training and testing is a job done before every moedel training
So this function does exactly that
}
\examples{
test_train_split(mtcars)
test_train_split(mtcars, train_size=0.75, seed=101, shuffle=TRUE)
}
\seealso{
\code{\link[base:grep]{base::gsub()}} which this function wraps.
}
