% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZ.R
\name{train_test_split}
\alias{train_test_split}
\title{Split dataset into train and test}
\usage{
train_test_split(data, train_size = 0.8, seed = 1, shuffle = FALSE)
}
\arguments{
\item{data}{Dataframes,Arrays,Vectors}

\item{train_size}{Numeric (should be between 0.0 and 1.0). Represent the proportion of the dataset to include in the test split.}

\item{seed}{Controls the shuffling applied to the data before applying the split. Pass an int for reproducible output across multiple function calls.}

\item{shuffle}{Whether or not to shuffle the data before splitting.}
}
\value{
List containing train-test split of inputs.
}
\description{
Splitting dataset into training and testing is a job done before every moedel training
So this function does exactly that
}
\seealso{
\code{\link[base:grep]{base::gsub()}} which this function wraps.
}
