\name{ez-internal}
\Rdversion{1.1}
\alias{ez-internal}
\title{Internal ez Functions}
\description{
  Internal functions for the \code{ez} package.
}
\details{
  These are not to be called by the user.
}
\author{Michael A Lawrence \email{Mike.Lawrence@dal.ca}}
\seealso{
	\code{\link{ANT}}, \code{\link{ANT2}}, \code{\link{ezANOVA}}, \code{\link{ezBoot}}, \code{\link{ezCor}}, \code{\link{ezDesign}}, \code{\link{ezPerm}}, \code{\link{ezPlot}}, \code{\link{ezPlotBoot}}, \code{\link{ezPrecis}}, \code{\link{ezStats}}
}
\keyword{ internal }
