\name{ez-internal}
\alias{ez-internal}
\title{Internal ez Functions}
\description{
Internal functions for the \code{ez} package.
}
\details{
These are not to be called by the user.
}
\author{
Michael A. Lawrence \email{Mike.Lawrence@dal.ca}
To report bugs or request features, please visit: \url{https://github.com/mike-lawrence/ez/issues}
To keep up to date on developments related to this package, join the discussion group at: \url{http://groups.google.com/group/ez4r}
}
\keyword{ internal }
