% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZ.R
\name{train_test_split}
\alias{train_test_split}
\title{Split dataset into train and test}
\usage{
train_test_split(data, train_size = 0.8, seed = 1, shuffle = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{train_size}{List of unwanted characters/ Identifier}

\item{seed}{Character to substitute/ Separator}

\item{ushuffel}{Boolean whether to use the default identifier list}
}
\value{
Dataframe with removed characters from the column names
}
\description{
Sometimes we may get a dataframe wich have column names have characters like !,&,(,),/,.,\link{,}," "
These are unwanted boviously coz they maek it impossible to reference the column using $ operator or while pipelining
So this function removes those characters from the column names.
}
\examples{
clean_names(data=df,idf=c("|","!","$"),sep='.',use.def=FALSE)
clean_names(data=df)
clean_names(data=df,sep='.')
}
\seealso{
\code{\link[base:grep]{base::gsub()}} which this function wraps.
}
